backend/requirements.txt
fastapi
uvicorn
chromadb
requests
pypdf
redis
psycopg2-binary
python-dotenv
